feature: refactor + adding repository layer.
