+++ app/src/main/java/com/mfpe/vinilos/data/model/Album.kt
@@ -1,4 +1,4 @@
-package com.mfpe.vinilos.models
+package com.mfpe.vinilos.data.model
 
 import java.io.Serializable
 import java.util.Date
