+++ app/src/main/java/com/mfpe/vinilos/viewmodel/AlbumListViewModel.kt
@@ -1,31 +1,30 @@
-package com.mfpe.vinilos.viewmodels
+package com.mfpe.vinilos.viewmodel
 
 import android.util.Log
 import androidx.lifecycle.LiveData
 import androidx.lifecycle.MutableLiveData
 import androidx.lifecycle.ViewModel
-import com.mfpe.vinilos.models.Album
-import com.mfpe.vinilos.network.AlbumService
-import com.mfpe.vinilos.network.RetrofitApiClient
+import com.mfpe.vinilos.data.model.Album
+import com.mfpe.vinilos.data.repository.AlbumRepository
 import retrofit2.Call
 import retrofit2.Callback
 import retrofit2.Response
 
 class AlbumListViewModel : ViewModel() {
 
-    private val albumService = RetrofitApiClient.createRetrofitService(AlbumService::class.java)
+    private val albumRepository: AlbumRepository = AlbumRepository()
 
     private val _albums = MutableLiveData<List<Album>>()
     val albums: LiveData<List<Album>> get() = _albums
 
-    private var _networkError = MutableLiveData<Boolean>(false)
+    private var _networkError = MutableLiveData(false)
     val networkError: LiveData<Boolean> get() = _networkError
 
     fun fetchAlbums() {
-        albumService.getAlbums().enqueue(object : Callback<List<Album>> {
-            override fun onResponse(call: Call<List<Album>>, response: Response<List<Album>>) {
-                if (response.isSuccessful) {
-                    _albums.value = response.body()
+        albumRepository.getAlbums().enqueue(object : Callback<List<Album>> {
+            override fun onResponse(call: Call<List<Album>>, res: Response<List<Album>>) {
+                if (res.isSuccessful) {
+                    _albums.value = res.body()
                     _networkError.value = false
                 }
             }
