+++ app/src/main/java/com/mfpe/vinilos/viewmodel/ArtistViewModel.kt
@@ -1,13 +1,13 @@
-package com.mfpe.vinilos.viewmodels
+package com.mfpe.vinilos.viewmodel
 
 import androidx.lifecycle.LiveData
 import androidx.lifecycle.MutableLiveData
 import androidx.lifecycle.ViewModel
 
-class CollectorViewModel : ViewModel() {
+class ArtistViewModel : ViewModel() {
 
     private val _text = MutableLiveData<String>().apply {
-        value = "This is notifications Fragment"
+        value = "This is dashboard Fragment"
     }
     val text: LiveData<String> = _text
 }
