+++ app/src/main/java/com/mfpe/vinilos/viewmodel/CollectorViewModel.kt
@@ -1,13 +1,13 @@
-package com.mfpe.vinilos.viewmodels
+package com.mfpe.vinilos.viewmodel
 
 import androidx.lifecycle.LiveData
 import androidx.lifecycle.MutableLiveData
 import androidx.lifecycle.ViewModel
 
-class ArtistViewModel : ViewModel() {
+class CollectorViewModel : ViewModel() {
 
     private val _text = MutableLiveData<String>().apply {
-        value = "This is dashboard Fragment"
+        value = "This is notifications Fragment"
     }
     val text: LiveData<String> = _text
 }
